==> docs/nlp/notebooks/12 Advanced -- Word2Vec.ipynb <==
from setup import *
import sys
if DATA_PATH not in sys.path: sys.path.append(DATA_PATH)
from constants import *

%matplotlib inline
display(HTML("<style>.container { width:100% !important; }</style>"))
pd.set_option('display.max_rows', 6)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 800)
pd.set_option('precision', 2)
%precision 4
%pprint

from gensim.models import Word2Vec

For now let's use the Google News model (300 topics)

wv = Word2Vec.load_word2vec_format(os.path.join(DATA_PATH, 'GoogleNews-vectors-negative300.bin.gz'), binary=True)


wv.most_similar(positive=['python', 'snake',], negative=['programming'])

wv.most_similar(positive=['PyCon'])

wv.most_similar(positive=['Portland', 'Oregon', ], negative=['city', 'government'])
